+++ docs/changelog.md
@@ -14,6 +14,7 @@ category: Administration
 * feat: configure the sorting order of user and project templates
 * docs: include configuration for sorting order
 * refactor: optimize i18n bundle for internationalization
+* test: add tests for configuration action
 
 ### [23.09.0] - 2023-09-01